+++ cat-home/src/main/java/com/dianping/cat/report/page/statistics/service/UtilizationReportService.java
@@ -81,7 +81,7 @@ public class UtilizationReportService extends AbstractReportService<UtilizationR
 		return utilizationReport;
 	}
 
-	private UtilizationReport queryFromDailyBinary(int id, String domain) {
+	private UtilizationReport queryFromDailyBinary(long id, String domain) {
 		DailyReportContent content = m_dailyReportContentDao.findByPK(id);
 
 		if (content != null) {
